Rigging3D/pyproject.toml
[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "rigging3d"
version = "0.5.0"
description = "CLI de auto-rigging 3D"
readme = "README.md"
requires-python = ">=3.10"
license = "MIT"
authors = [{ name = "Rigging3D Project" }]
dependencies = [
    "gamedev-shared @ file:../Shared",
    "click>=8.1.0",
    "rich>=13.0.0",
    "rich-click>=1.8.0",
    "PyYAML>=6.0",
]

[project.optional-dependencies]
inference = [
    "torch>=2.3.1",
    "torchvision>=0.18.0",
    "numpy>=1.26.4,<3",
    "transformers==4.51.3",
    "python-box",
    "einops",
    "omegaconf",
    "pytorch_lightning",
    "lightning",
    "addict",
    "timm",
    "fast-simplification",
    "bpy==4.2.0",     # wheels PyPI só cp311 — usar Python 3.11
    "trimesh",
    "open3d",
    "pyrender",
    "huggingface_hub",
    "wandb",
    # Deps CUDA-specific (instaladas por setup.sh / installer.py --inference):
    #   spconv-cu121, cumm-cu121, torch-scatter, torch-cluster
    # flash_attn: opcional (fallback SDPA). Instalar com: bash scripts/install_flash_attn.sh
]
dev = ["pytest>=7.4.0", "pytest-cov>=4.0.0", "ruff>=0.11.0"]

[project.scripts]
rigging3d = "rigging3d.cli:main"

[tool.setuptools]
include-package-data = true

[tool.setuptools.packages.find]
where = ["src"]

[tool.setuptools.package-data]
rigging3d = ["unirig/**/*"]

[tool.pytest.ini_options]
pythonpath = ["src"]
testpaths = ["tests"]

[tool.ruff]
line-length = 120
src = ["src/rigging3d"]
exclude = ["src/rigging3d/unirig"]

[tool.ruff.lint]
select = ["E", "F", "I", "UP", "W"]

[tool.ruff.lint.per-file-ignores]
# Import após sys.path.insert (execução directa sem pip install do Shared)
"scripts/installer.py" = ["E402"]
